server/app/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0
authlib>=1.2.0
python-jose>=3.3.0
pydantic>=2.6.0
requests>=2.31.0